// admin/src/api/generated/schemas/adminCreateMenuRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */

export type AdminCreateMenuRequest = {
  name?: string;
  pid?: string;
  path?: string;
  icon?: string;
  hide_in_menu?: boolean;
  hide_children_in_menu?: boolean;
  sort?: number;
  remark?: string;
  status?: number;
  target?: string;
  access?: string;
  key?: string;
};